🥅 ♻️ consider none match pipeline as error
Before the change `kratix build` would panic
when the defined pipeline name in the `promise.yaml`
does not match the folder name.

The refactoring provides an error messages to state the assumption and
avoids checking for negative indices in higher order functions
which leads higher resilience and less code being written.